Add files via upload
Updated version of the MakeChange program. This program uses the same calculations as MakeChange, but calculates change from user input. User input is received from a string method named AskUser. AskUser prints its parameter and reads user input as a string. 

The AskUser method is called in the main method to get user input for the amount they are spending on a candy bar, and how much that candy bar costs. This input is then converted from a string to a double so that the change in coins can be calculated.

This program is an exercise in creating methods, calling methods, and data type conversion.