Add files via upload
This program is similar to the MakeChange2 program listed in this repository. A string method named AskUser is created to print the method's parameter and receive user input.

The AskUser method is called in the main method to receive user input for coefficients a, b, and c which are used in the quadratic equation. This program introduces the concept of conditional statements in the form of if statements. 

The roots of a quadratic function cannot be calculated if coefficient a == 0 or if the discriminant (b^2-4ac) is less than 0. Another conditional is used to calculate the x1 root if the discriminant is equal to 0. 

Otherwise, the roots x1 and x2 are calculated and printed. 

This program is in an exercise in creating a method, calling a method, and using if statements (conditionals).